requirements.txt
Django==5.0.4
mysqlclient==2.2.4    
python-decouple==3.8    
Pillow==10.3.0          
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.14.0
pandas==2.1.1
openpyxl==3.1.2
PyJWT==2.8.0
python-dateutil==2.8.2
